requirements.txt
mcp[cli]>=1.2.0
anthropic>=0.45.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
